backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
yfinance==1.2.0
pydantic==2.10.3
httpx==0.27.2
boto3==1.35.81
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
psycopg2-binary==2.9.10
yahooquery
